started implementation of prescription component as cards
